+++ modules/entity/api/PhotoObject.php
@@ -14,13 +14,13 @@ class PhotoObject extends \yii\easyii\components\ApiObject
         $img = Html::img($this->thumb($width, $height));
         $a = Html::a($img, $this->image, [
             'class' => 'easyii-box',
-            'rel' => 'catalog-'.$this->model->item_id,
+            'rel' => 'entity-'.$this->model->item_id,
             'title' => $this->description
         ]);
         return LIVE_EDIT ? API::liveEdit($a, $this->editLink) : $a;
     }
 
     public function getEditLink(){
-        return Url::to(['/admin/catalog/items/photos', 'id' => $this->model->item_id]).'#photo-'.$this->id;
+        return Url::to(['/admin/entity/items/photos', 'id' => $this->model->item_id]).'#photo-'.$this->id;
     }
 }